0715게시판 업그레이드

+++ spring06integrated/src/main/java/com/kh/spring06/controller/BoardController.java
@@ -13,6 +13,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.kh.spring06.dao.BoardDao;
 import com.kh.spring06.dto.BoardDto;
+import com.kh.spring06.error.TargetNotFoundException;
+
+import jakarta.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/board")
@@ -44,19 +47,25 @@ public class BoardController {
 		return "/WEB-INF/views/board/write.jsp";
 	}
 	@PostMapping("/write")
-	public String write(@ModelAttribute BoardDto boardDto) {
+	public String write(@ModelAttribute BoardDto boardDto, HttpSession session) {
+		String createdUser = (String)session.getAttribute("createdUser");
+		//시퀀스 번호를 먼저 생성하도록 지시
+		int seq = boardDao.sequence();
+		boardDto.setBoardWriter(createdUser);
+		boardDto.setBoardNo(seq);
 		boardDao.write(boardDto);
-		return "redirect:writeFinish";
+		return "redirect:read?boardNo=" + seq;
 	}
-	@RequestMapping("/writeFinish")
-	public String writeFinish() {
-		return "/WEB-INF/views/board/writeFinish.jsp";
-	}
+//	@RequestMapping("/writeFinish")
+//	public String writeFinish() {
+//		return "/WEB-INF/views/board/writeFinish.jsp";
+//	}
 	
 	//읽기
 	@RequestMapping("/read")
 	public String read(@RequestParam int boardNo, Model model) {
 		BoardDto boardDto = boardDao.selectOne(boardNo);
+		if(boardDto == null) throw new TargetNotFoundException("존재하지 않는 글 번호");
 		model.addAttribute("boardDto", boardDto);
 		return "/WEB-INF/views/board/read.jsp";
 	}
@@ -65,19 +74,24 @@ public class BoardController {
 	@GetMapping("/update")
 	public String update(Model model, @RequestParam int boardNo) {
 		BoardDto boardDto = boardDao.selectOne(boardNo);
+		if(boardDto == null) throw new TargetNotFoundException("존재하지 않는 글 번호");
 		model.addAttribute("boardDto", boardDto);
 		return "/WEB-INF/views/board/update.jsp";
 	}
 	
 	@PostMapping("/update")
 	public String update(@ModelAttribute BoardDto boardDto) {
+		if(boardDto == null) throw new TargetNotFoundException("존재하지 않는 글 번호");
 		boardDao.update(boardDto);
 		return "redirect:read?boardNo="+boardDto.getBoardNo();
 	}
 	
 	@RequestMapping("/delete")
 	public String delete(@RequestParam int boardNo) {
+		BoardDto boardDto = boardDao.selectOne(boardNo);
+		if(boardDto == null) throw new TargetNotFoundException("존재하지 않는 글 번호");
 		boardDao.delete(boardNo);
 		return "redirect:list";
 	}
 }
+
